# src/main/java/com/juansenen/gaticket/doc/Gaticket_API_doc.yaml
openapi: 3.0.1
info:
  title: GATICKET API
  version: "1.0.0"
  description: API diseñada para el control y consulta de la web app Gaticket. Gestor de incidencias técnicas, al cual pueden conectarse los usuarios para poder comunicar al equipo de Tecnologias de la Información de la empresa las incidencias surgidas en sus equipos; y por otra parte los administradores pueden recibirlas, gestionarlas y llevar un control sobre la base de datos de los dispositivos a su cargo.

    Para poder realizar un monitoreo y control de la API, se ha incuido la libreria spring-doc actuator. Accesible en el siguiente link

    Todos los endppoints de actuator [API actuator](http://localhost:8080/actuator)
    endpoints de metricas [API actuator](http://localhost:8080/actuator/metrics)
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  /departments:
    get:
      tags:
        - department
      summary: Retrieve all departments
      description: Get all departments on the data base.
      operationId: getCollectionResource-department-get_1_1
      responses:
        "400":
          description: Invalid
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelDepartment'
            application/x-spring-data-compact+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelDepartment'
            text/uri-list:
              schema:
                type: string
    post:
      tags:
        - department
      summary: Save a new department
      description: 'Save one department '
      operationId: postCollectionResource-department-post_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
        required: true
      responses:
        "400":
          description: Invalid
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelDepartment'
  /departments/{id}:
    get:
      tags:
        - department
      summary: Retrieve a department by Id
      description: Get one department knowing his Id.
      operationId: getItemResource-department-get_1
      parameters:
        - name: id
          in: path
          description: ID of department search
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Invalid
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not found
          content: {}
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelDepartment'
    put:
      tags:
        - department
      summary: Update a department by Id
      description: Update one department knowing his Id.
      operationId: putItemResource-department-put_1
      parameters:
        - name: id
          in: path
          description: Department Id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
        required: true
      responses:
        "400":
          description: Invalid
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelDepartment'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelDepartment'
        "204":
          description: No Content
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
    delete:
      tags:
        - department
      summary: Delete a department by Id
      description: Delete one department knowing his Id.
      operationId: deleteItemResource-department-delete_1
      parameters:
        - name: id
          in: path
          description: Id of the department to erase
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Not found
          content: {}
        "204":
          description: No Content
        "201":
          description: Created
...
  /history:
    get:
      tags:
        - incidenceHistory
      summary: Retrieve all incidences history
      description: Get all incidences saved like history on data base.
      operationId: getAllHistory
      responses:
        "400":
          description: Invalid
          content: {}
        "404":
          description: Resource not found
          content: {}
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incidences'
        "500":
          description: Server error
    post:
      tags:
        - incidenceHistory
      summary: Save a incidence solution to history
      description: Save a incidence solution to history
      operationId: saveToHistory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidencesHistory'
        required: true
      responses:
        "400":
          description: Invalid
          content: {}
        "404":
          description: Bad request
          content: {}
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incidences'
        "500":
          description: Server error
...
  /department/{iduser}:
    get:
      tags:
        - department
      summary: Retrieve a department by Id user
      description: Get one department knowing id of a user.
      operationId: findDepartUser
      parameters:
        - name: iduser
          in: path
          description: ID of user search
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid
          content: {}
        "404":
          description: Not found
          content: {}
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
components:
  schemas:
    ErrorMessage:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
...
    IncidencesHistory:
      type: object
      properties:
        idhistory:
          type: integer
          description: history Id
          format: int64
          readOnly: true
          example: 12
        historyTip:
          type: string
          description: TIP of the user who record incidence
          example: T45678I
        historyTheme:
          type: string
          description: Theme of the history incidence
          example: Mi pc se reinicia solo
        historyCommit:
          type: string
          description: Commit of the history incidence
          example: Cada cierto tiempo el pc se reinica sin previo aviso
        historyDateFinish:
          type: string
          description: Date when history incidence was fixit
          format: date-time
        historyAdmin:
          type: string
          description: Admin who solved incidence
          example: A58656F
        historySolution:
          type: string
          description: Solution to incidence
          example: User must to restar the computer and admin must to change assigned
            IP
...
    Department:
      type: object
      properties:
        departemtId:
          type: integer
          description: DepartmentId
          format: int64
          readOnly: true
          example: 6
        departmentName:
          type: string
          description: Department name
          example: sales
        departmentPhone:
          type: string
          description: Department contact phone
          example: "0700047"
        departmentMail:
          type: string
          description: Department contact mail
          example: sales@megacompany.com
        departmentCity:
          type: string
          description: City where it is the department
          example: Zaragoza
        devices:
          type: array
          items:
            $ref: '#/components/schemas/Device'
        userList:
          type: array
          items:
            $ref: '#/components/schemas/User'
      description: Departments inside de company
    Device:
      type: object
      properties:
        deviceId:
          type: integer
          description: DeviceId
          format: int64
          readOnly: true
          example: 12
        deviceHd:
          type: integer
          description: Device HD capacity
          format: int32
          example: 500
        deviceRam:
          type: integer
          description: Device RAM memory installed
          format: int32
          example: 32
        deviceMac:
          maxLength: 17
          minLength: 12
          type: string
          description: Device mac identification
          example: 13:BG:45:89:99:AC
        deviceSerial:
          type: string
          description: Device serial number
          example: CXV5678D
        deviceModel:
          type: string
          description: Device model
          example: Acer TravelMate
        deviceDateBuy:
          type: string
          description: Device date of buy MM/dd/yyyy
          format: date-time
        deviceDateStart:
          type: string
          description: Device date begin to work or installed
          format: date-time
        deviceType:
          $ref: '#/components/schemas/Type'
        net:
          $ref: '#/components/schemas/Net'
      description: All class of devices
    Net:
      type: object
      properties:
        netId:
          type: integer
          description: netId
          format: int64
          readOnly: true
          example: 8
        netIp:
          type: string
          description: IP
          example: 10.53.3.456
        netMask:
          type: string
          description: Mask of network
          example: 255.255.255.0
        netCdir:
          type: string
          description: CDIR number
          example: "24"
        netGateWay:
          type: string
          description: Gateway number
          example: 10.52.2.1
        netStatus:
          type: boolean
          description: Status of the IP
          example: false
      description: Network information of a device
    Type:
      required:
        - typeName
      type: object
      properties:
        typeId:
          type: integer
          description: TypeId
          format: int64
          readOnly: true
          example: 4
        typeName:
          type: string
          description: Type of device
          example: Laptop
      description: Type of device information
    User:
      required:
        - userMail
        - userPassword
        - userRol
        - userTip
      type: object
      properties:
        userId:
          type: integer
          description: UserId
          format: int64
          readOnly: true
          example: 2
        userMail:
          type: string
          description: User email
          example: juanito@gmail.com
        userPassword:
          type: string
          description: User password
          example: xXFG7ujka123
        userTip:
          type: string
          description: User personal number identification
          example: A58656F
        userRol:
          type: string
          description: User rol permision on the API
          example: usuario
      description: User information
    EntityModelDepartment:
      type: object
      properties:
        departmentName:
          type: string
          description: Department name
          example: sales
        departmentPhone:
          type: string
          description: Department contact phone
          example: "0700047"
        departmentMail:
          type: string
          description: Department contact mail
          example: sales@megacompany.com
        departmentCity:
          type: string
          description: City where it is the department
          example: Zaragoza
        _links:
          $ref: '#/components/schemas/Links'
    CollectionModelEntityModelDepartment:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            departments:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelDepartment'
        _links:
          $ref: '#/components/schemas/Links'
...
    Incidences:
      type: object
      properties:
        incidencesId:
          type: integer
          description: incidenceId
          format: int64
          readOnly: true
          example: 12
        incidenceCommit:
          type: string
          description: Description
          example: My computer has began to go slowly. And I can´t see mp4 videos
        incidenceStatus:
          type: string
          description: Incidence status
          example: Active
        incidenceDate:
          type: string
          description: Incidence date start
          format: date-time
        incidenceDateFinish:
          type: string
          description: Incidence date fix it
          format: date-time
        incidenceTheme:
          type: string
          description: Incidence theme
          example: Computer damage
        adminId:
          type: integer
          description: Admin ID who fix incidence
          format: int32
        device:
          $ref: '#/components/schemas/Device'
        user:
          $ref: '#/components/schemas/User'
        responsable:
          $ref: '#/components/schemas/User'
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean